requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
transformers==4.38.1
torch==2.2.0
Pillow==10.2.0
httpx==0.26.0
python-dotenv==1.0.1
numpy==1.26.4
google-generativeai==0.8.5 